refactor(acceptance): extract scenario 20 helpers into sibling file
Move the file-local helper group (fixture builder, run/preview polling,
event fetch, proxy/login handshake) out of 20-preview.ts into a new
20-preview.helpers.ts sibling, mirroring 32-plot-workbench-loop.helpers.ts.
Both files now sit under the 500-line budget (368 / 320).

warren-65f6
